refactor: move unzip to python package
